Affects Whole Project
Relocates files to create a more self-contained package

All files in custom_templates have been relocated to custom_templates\zenos_ai and all references to import those templates have been updated to look in that subfolder. This reduces the possibility of overwriting user's existing files with the same name and also namespaces the jinja used by zenos.

The following folders are relocated (verbatim) from the root to packages/zenos_ai/xxx
- assets
- automations
- docs
- scripts

The scripts and automations have been updated to start with the expected dictionary key as required for their presence in a package.
- script:
- automation:

All template sensors have been relocated to packages/zenos_ai/sensors (in keeping with the naming conventions already in use).